[Dash] Always run `Modify Info.plist` and `Embed Mapbox Access Token`
Run script build phase 'Modify Info.plist' will be run during every build because it does not specify any outputs. To address this warning, either add output dependencies to the script phase, or configure it to run in every build by unchecking "Based on dependency analysis" in the script phase.